Moved gstreamer code to portable
Made a miro.frontends.widgets.gst package that contains portable gstreamer
code.  Split up the renderer and sniffer code.

Made a portable make_file_url() method, that works on both linux and windows.
This was almost the only difference between the gstreamer code on both
platforms.

The gst package still has some GTK-specific bits to it.  I plan to remove
these soon.
