+++ tv/linux/setup.py
@@ -567,6 +567,7 @@ setup(name='miro',
         'miro.frontends.profilewidgets',
         'miro.frontends.shell',
         'miro.frontends.widgets',
+        'miro.frontends.widgets.gst',
         'miro.frontends.widgets.gtk',
         'miro.plat',
         'miro.plat.frontends',
